requirements.txt
# NMBS Train Data API Requirements
gtfs-realtime-bindings>=1.0.0
protobuf>=3.20.0
beautifulsoup4>=4.10.0
requests>=2.27.0
python-dotenv>=1.0.0
cloudscraper>=1.2.71
schedule>=1.1.0
flask>=2.0.0
flask-cors>=3.0.10
lxml>=4.9.0
charset-normalizer
colorama>=0.4.6
pandas>=1.3.0